Updated the UnitTests so they can call the methods in main.py
